requirements.txt
pre-commit==3.5.0
numpy==1.26.2
pandas==2.1.3
langchain==0.0.339
dash[diskcache]==2.14.1
dash-bootstrap-components==1.5.0
python-dotenv==1.0.0
